changed to long double

+++ source/ldl_driver.cpp
@@ -68,7 +68,7 @@ int main(int argc, char* argv[])
 		return 0;
 	}
 
-	sparse_ainv::solver<double> solv;
+	sparse_ainv::solver<long double> solv;
 
 	//default is statistics output
 	solv.set_message_level("statistics");
@@ -90,7 +90,7 @@ int main(int argc, char* argv[])
 			FLAGS_max_iters = 200;
 		}
 
-		vector<double> rhs;
+		vector<long double> rhs;
 		if (!FLAGS_rhs_file.empty()) {
 			sparse_ainv::read_vector(rhs, FLAGS_rhs_file);
 		} else {
